requirements.txt
fastapi
uvicorn[standard]
discord.py
pdfplumber
tabulate
python-multipart
requests 
pandas 
python-dotenv 
pandasql
langchain 
langchain-community 
langchain-openai